Network.Google.Data.Time

Description

Documentation

data Time'

This SHOULD be a time in the format of hh:mm:ss. It is recommended that you use the "date-time" format instead of "time" unless you need to transfer only the time part.

data Date'

This SHOULD be a date in the format of YYYY-MM-DD. It is recommended that you use the "date-time" format instead of "date" unless you need to transfer only the date part.

newtype DateTime'

This SHOULD be a date in ISO 8601 format of YYYY-MM- DDThh:mm:ssZ in UTC time. This is the recommended form of date/timestamp.

Constructors

DateTime'

Fields

unDateTime :: UTCTime

data GDuration

A duration in seconds with up to nine fractional digits, terminated by s.

Example: "3.5s".

_Time :: Iso' Time' TimeOfDay

_Date :: Iso' Date' Day

_DateTime :: Iso' DateTime' UTCTime

_GDuration :: Iso' GDuration Scientific