Network.Google.Types

Description

Documentation

data AltJSON

Constructors

AltJSON

data AltMedia

Constructors

AltMedia

data Multipart

Constructors

Multipart

newtype OAuthScope

An OAuth2 scope.

Constructors

OAuthScope Text

newtype AccessToken

An OAuth2 access token.

Constructors

AccessToken Text

newtype RefreshToken

An OAuth2 refresh token.

Constructors

RefreshToken Text

newtype ClientId

A client identifier.

Constructors

ClientId Text

newtype ServiceId

A service identifier.

Constructors

ServiceId Text

newtype GSecret

An opaque client secret.

Constructors

GSecret Text

newtype MediaDownload a

Constructors

MediaDownload a

data MediaUpload a

Constructors

MediaUpload a GBody

_Coerce :: (Coercible a b, Coercible b a) => Iso' a b

_Default :: Monoid a => Iso' (Maybe a) a

Invalid Iso, exists for ease of composition with the current 'Lens . Iso' chaining to hide internal types from the user.

type Stream = ConduitM () ByteString (ResourceT IO) ()

data Error

Constructors

TransportError HttpException

SerializeError SerializeError

ServiceError ServiceError

data SerializeError

Constructors

SerializeError'

Fields

_serializeId :: !ServiceId

_serializeHeaders :: [Header]

_serializeStatus :: !Status

_serializeMessage :: !String

_serializeBody :: !(Maybe ByteString)

data ServiceError

Constructors

ServiceError'

Fields

_serviceId :: !ServiceId

_serviceStatus :: !Status

_serviceHeaders :: ![Header]

_serviceBody :: !(Maybe ByteString)

class AsError a where

Minimal complete definition

_Error

Methods

_Error :: Prism' a Error

A general Amazonka error.

_TransportError :: Prism' a HttpException

An error occured while communicating over HTTP with a remote service.

_SerializeError :: Prism' a SerializeError

A serialisation error occured when attempting to deserialise a response.

_ServiceError :: Prism' a ServiceError

A service specific error returned by the remote service.

data ServiceConfig

Constructors

ServiceConfig

Fields

_svcId :: !ServiceId

_svcHost :: !ByteString

_svcPath :: !Builder

_svcPort :: !Int

_svcSecure :: !Bool

_svcTimeout :: !(Maybe Seconds)

defaultService :: ServiceId -> ByteString -> ServiceConfig

serviceHost :: Lens' ServiceConfig ByteString

The remote host name, used for both the IP address to connect to and the host request header.

servicePort :: Lens' ServiceConfig Int

The remote port to connect to.

Defaults to 443.

servicePath :: Lens' ServiceConfig Builder

A path prefix that is prepended to any sent HTTP request.

Defaults to mempty.

serviceSecure :: Lens' ServiceConfig Bool

Whether to use HTTPS/SSL.

Defaults to True.

serviceTimeout :: Lens' ServiceConfig (Maybe Seconds)

Number of seconds to wait for a response.

data GBody

A single part of a (potentially multipart) request body.

Note: The IsString instance defaults to a text/plain MIME type.

Constructors

GBody !MediaType !RequestBody

bodyContentType :: Lens' GBody MediaType

A lens into the MediaType of a request Body.

data Request

An intermediary request builder.

Constructors

Request

Fields

_rqPath :: !Builder

_rqQuery :: !(DList (ByteString, Maybe ByteString))

_rqHeaders :: !(DList (HeaderName, ByteString))

_rqBody :: ![GBody]

appendPath :: Request -> Builder -> Request

appendPaths :: ToHttpApiData a => Request -> [a] -> Request

appendQuery :: Request -> ByteString -> Maybe Text -> Request

appendHeader :: Request -> HeaderName -> Maybe Text -> Request

setBody :: Request -> [GBody] -> Request

data GClient a

A materialised 'http-client' request and associated response parser.

Constructors

GClient

Fields

_cliAccept :: !(Maybe MediaType)

_cliMethod :: !Method

_cliCheck :: !(Status -> Bool)

_cliService :: !ServiceConfig

_cliRequest :: !Request

_cliResponse :: !(Stream -> ResourceT IO (Either (String, ByteString) a))

clientService :: Lens' (GClient a) ServiceConfig

mime :: FromStream c a => Proxy c -> Method -> [Int] -> Request -> ServiceConfig -> GClient a

discard :: Method -> [Int] -> Request -> ServiceConfig -> GClient ()

gClient :: (Stream -> ResourceT IO (Either (String, ByteString) a)) -> Maybe MediaType -> Method -> [Int] -> Request -> ServiceConfig -> GClient a

class Accept c => ToBody c a where

Methods

toBody :: Proxy c -> a -> GBody

class Accept c => FromStream c a where

Methods

fromStream :: Proxy c -> Stream -> ResourceT IO (Either (String, ByteString) a)

class GoogleRequest a where

Associated Types

type Rs a :: *

type Scopes a :: [Symbol]

Methods

requestClient :: a -> GClient (Rs a)

class GoogleClient fn where

Associated Types

type Fn fn :: *

Methods

buildClient :: Proxy fn -> Request -> Fn fn

data Captures (s :: Symbol) a

Multiple path captures, with [xs] forming x1x2x2....

data CaptureMode (s :: Symbol) (m :: Symbol) a

Form a Google style sub-resource, such as /capture:mode.

data MultipartRelated (cs :: [*]) m

sinkLBS :: Stream -> ResourceT IO ByteString

buildText :: ToHttpApiData a => a -> Builder

buildSymbol :: forall n proxy. KnownSymbol n => proxy n -> Builder

byteSymbol :: forall n proxy. KnownSymbol n => proxy n -> ByteString

newtype Seconds

An integral value representing seconds.

Constructors

Seconds Int

seconds :: Seconds -> Int

microseconds :: Seconds -> Int

newtype GFieldMask

Constructors

GFieldMask Text